Move all the CLI enum definitions centrally (to avoid chain imports)
These enums are defined primarily for use in defining the CLI with typer, so to
me putting them in public_cli_args.py is logical.
