+++ pyani_plus/methods/anim.py
@@ -41,18 +41,11 @@ percentage (of whole genome) for each pairwise comparison.
 """
 
 from collections import defaultdict
-from enum import Enum
 from pathlib import Path
 
 import intervaltree  # type: ignore  # noqa: PGH003
 
-
-class EnumModeANIm(str, Enum):
-    """Enum for the --mode command line argument passed to ANIm."""
-
-    mum = "mum"  # default
-    maxmatch = "maxmatch"
-
+from pyani_plus.public_cli_args import EnumModeANIm
 
 MODE = EnumModeANIm.mum  # constant for CLI default
